refactor: relocate provider as fsd
