+++ packages/client/tsconfig.json
@@ -32,5 +32,5 @@
       }
    },
    "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
-   "exclude": ["node_modules"]
+   "exclude": ["node_modules", ".next", "dist"]
 }
